Move Fields to their own separate directory
