+++ src/Fields/DurationRange.js
@@ -1,5 +1,6 @@
+import Immutable  from '../Types/Immutable';
+
 import CustomDate from './CustomDate';
-import Immutable  from './Immutable';
 
 /** Class to keep duration ranges
  *
