+++ src/Streams/Balance.js
@@ -1,6 +1,6 @@
 import { first, map, share } from 'rxjs/operators';
 
-import Monetary              from '../Types/Monetary';
+import Monetary              from '../Fields/Monetary';
 import Stream                from '../Types/Stream';
 
 /**
